refactor: useDebounce, product 분리
